Move representation tests to jdk11
