+++ src/jvm/test-jdk11/org/jetbrains/kotlinx/lincheck_test/representation/VariableReadWriteRunWithLambdaTest.kt
